Indexer almost works
