+++ package.json
@@ -84,6 +84,7 @@
     "storybook": "^8.3.6",
     "storybook-dark-mode": "^4.0.2",
     "storybook-preset-inline-svg": "^1.0.1",
+    "ts-dedent": "^2.2.0",
     "typescript": "^5.6.3",
     "vite": "^5.4.10",
     "vite-plugin-svgr": "^4.3.0",
